+++ MoltenVK/MoltenVK/GPUObjects/MVKQueue.mm
@@ -360,19 +360,18 @@ id<MTLCommandBuffer> MVKQueuePresentSurfaceSubmission::getMTLCommandBuffer() {
 
 MVKQueuePresentSurfaceSubmission::MVKQueuePresentSurfaceSubmission(MVKQueue* queue,
 																   const VkPresentInfoKHR* pPresentInfo)
-		: MVKQueueSubmission(queue, pPresentInfo->waitSemaphoreCount, pPresentInfo->pWaitSemaphores) {
+	: MVKQueueSubmission(queue, pPresentInfo->waitSemaphoreCount, pPresentInfo->pWaitSemaphores) {
 
-	// Populate the array of swapchain images, testing each one for a change in surface size
-	_presentableImages.reserve(pPresentInfo->swapchainCount);
-	for (uint32_t i = 0; i < pPresentInfo->swapchainCount; i++) {
-		MVKSwapchain* mvkSC = (MVKSwapchain*)pPresentInfo->pSwapchains[i];
-		_presentableImages.push_back(mvkSC->getPresentableImage(pPresentInfo->pImageIndices[i]));
-		// Surface loss takes precedence over out-of-date errors.
-		if (mvkSC->getIsSurfaceLost()) {
-			setConfigurationResult(VK_ERROR_SURFACE_LOST_KHR);
-		} else if (mvkSC->getHasSurfaceSizeChanged() && getConfigurationResult() != VK_ERROR_SURFACE_LOST_KHR) {
-			setConfigurationResult(VK_ERROR_OUT_OF_DATE_KHR);
-		}
+	// Populate the array of swapchain images, testing each one for status
+	uint32_t scCnt = pPresentInfo->swapchainCount;
+	VkResult* pSCRslts = pPresentInfo->pResults;
+	_presentableImages.reserve(scCnt);
+	for (uint32_t scIdx = 0; scIdx < scCnt; scIdx++) {
+		MVKSwapchain* mvkSC = (MVKSwapchain*)pPresentInfo->pSwapchains[scIdx];
+		_presentableImages.push_back(mvkSC->getPresentableImage(pPresentInfo->pImageIndices[scIdx]));
+		VkResult scRslt = mvkSC->getSurfaceStatus();
+		if (pSCRslts) { pSCRslts[scIdx] = scRslt; }
+		setConfigurationResult(scRslt);
 	}
 }